Add splitpipe.py
'splitpipe' is a tool for partititioning parts of Unix pipelines into
a number of different executions. That is, after every N lines
(specified as a parameter), the process that was launched using the
command line arguments passed to splitpipe will receive an EOF. After
the previous process has finished, a new process will be executed,
continuing from where the previous one left. New processes will be
spawned for every N lines of input until splitpipe receives an EOF.

The need for this originally arose from the tendency of 'most' NLP
software to halt or crash after processing large inputs. It was
cumbersome to store and process many small files in the file system,
and thus splitpipe was born.